feat: Refactor Appointment and Team services to use query assemblers for improved response handling and code clarity
