+++ src/main/java/com/example/tmta/service/TeamService.java
@@ -21,13 +21,11 @@ import com.example.tmta.repository.MemberRepository;
 import com.example.tmta.repository.TeamMembersRepository;
 import com.example.tmta.repository.TeamRepository;
 import com.example.tmta.security.CurrentMemberProvider;
+import com.example.tmta.service.assembler.TeamQueryAssembler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -42,6 +40,7 @@ public class TeamService {
     private final TeamMembersRepository teamMembersRepository;
     private final AppointmentRepository appointmentRepository;
     private final CurrentMemberProvider currentMemberProvider;
+    private final TeamQueryAssembler teamQueryAssembler;
 
     @Transactional(readOnly = true)
     public TeamListResponseDto getTeamList() {
@@ -67,30 +66,7 @@ public class TeamService {
         Map<Long, Member> memberMap = memberRepository.findAllById(memberIds).stream()
                 .collect(Collectors.toMap(Member::getId, member -> member));
 
-        List<TeamListResponseDto.TeamInfo> teamInfoList = new ArrayList<>();
-        for (TeamMembers myMembership : myMemberships) {
-            Team team = teamMap.get(myMembership.getTeamId());
-            if (team == null) {
-                continue;
-            }
-
-            List<TeamMembers> teamMembers = membersByTeam.getOrDefault(team.getId(), List.of());
-            List<Appointment> teamAppointments = appointmentsByTeam.getOrDefault(team.getId(), List.of());
-
-            List<MemberInfo> members = teamMembers.stream()
-                    .map(membership -> toMemberInfo(membership, memberMap.get(membership.getMemberId())))
-                    .toList();
-            teamInfoList.add(new TeamListResponseDto.TeamInfo(
-                    team.getId(),
-                    team.getName(),
-                    teamAppointments.isEmpty() ? AppointmentState.CREATING : teamAppointments.get(0).getState(),
-                    (long) teamMembers.size(),
-                    members,
-                    myMembership.isTeamProfileSetupCompleted()
-            ));
-        }
-
-        return new TeamListResponseDto(teamInfoList);
+        return teamQueryAssembler.toTeamListResponse(myMemberships, teamMap, membersByTeam, appointmentsByTeam, memberMap);
     }
 
     @Transactional
@@ -124,23 +100,8 @@ public class TeamService {
         Map<Long, Member> memberMap = memberRepository.findAllById(memberIds).stream()
                 .collect(Collectors.toMap(Member::getId, member -> member));
 
-        List<MemberInfo> members = memberships.stream()
-                .map(membership -> toMemberInfo(membership, memberMap.get(membership.getMemberId())))
-                .toList();
-
         List<Appointment> appointments = appointmentRepository.findAllByTeamId(teamId);
-        List<DetailTeamList.AppointmentDetail> appointmentDetails = appointments.stream()
-                .map(appointment -> toAppointmentDetail(appointment, memberships.size()))
-                .toList();
-
-        return new DetailTeamList(
-                team.getId(),
-                team.getName(),
-                (long) memberships.size(),
-                team.getProfileImage(),
-                appointmentDetails,
-                members
-        );
+        return teamQueryAssembler.toDetailTeamList(team, memberships, appointments, memberMap);
     }
 
     @Transactional
@@ -230,39 +191,6 @@ public class TeamService {
         if (membership.getTeamRole() != TeamRole.ADMIN) {
             throw new BusinessException(ErrorCode.NOT_A_LEADER_OF_TEAM);
         }
-    }
-
-    private MemberInfo toMemberInfo(TeamMembers membership, Member member) {
-        if (member == null) {
-            return MemberInfo.of(membership.getMemberId(), null, null);
-        }
-        String displayName = membership.getTeamNickName() != null ? membership.getTeamNickName() : member.getNickName();
-        if (displayName == null) {
-            displayName = member.getName();
-        }
-        String profileImage = membership.getTeamProfileImage() != null ? membership.getTeamProfileImage() : member.getProfileImage();
-        return MemberInfo.of(member.getId(), displayName, profileImage);
-    }
-
-    private DetailTeamList.AppointmentDetail toAppointmentDetail(Appointment appointment, int memberCount) {
-        List<LocalDate> dates = appointment.getAppointmentDateList().stream()
-                .map(date -> date.getDate())
-                .sorted(Comparator.naturalOrder())
-                .toList();
-        LocalDate startDate = null;
-        LocalDate endDate = null;
-        if (!dates.isEmpty()) {
-            startDate = dates.get(0);
-            endDate = dates.get(dates.size() - 1);
-        }
-        return new DetailTeamList.AppointmentDetail(
-                appointment.getId(),
-                startDate,
-                endDate,
-                (long) memberCount,
-                appointment.getState(),
-                appointment.isOnlyDate()
-        );
     }
 
     private String normalizeTeamName(String teamName) {
